refactor: move error-line classifier to neutral ScpiResponseClassifier
Decouple DaqifiStreamingDevice from SdCardFileListParser by relocating
the shared IsErrorResponseLine/MatchesErrorPrefix helpers into a neutral
internal ScpiResponseClassifier class. Previously the general SCPI
response classifier reached into the SD-card-specific module, coupling
unrelated layers.

Also fix an OS-dependent test expectation: the FilenamesWithSingleBangSurvive
theory used Path.GetFileName, which treats '\\' as a separator on Windows
but not on Linux/macOS. Replaced with an explicit LastIndexOf('/') split
since the device protocol uses forward slashes.

No behavior change. All 902 tests pass on net9.0 + net10.0. Verified
on real device (Nq1, FW 3.4.4) via --sd-list — normal listings parse
correctly.
