+++ src/Daqifi.Core/Device/SdCard/SdCardFileListParser.cs
@@ -40,16 +40,11 @@ namespace Daqifi.Core.Device.SdCard
 
                 var path = line.Trim();
 
-                // Skip SCPI error responses: "**ERROR: -200, ..." or
-                // "ERROR: -200, ...". Bare "ERROR" prefix can't be used
-                // here because filenames like "error_log.csv" /
-                // "Errors_summary.bin" emitted without the Daqifi/
-                // directory prefix would also match (closes #190 second
-                // location — IsNonResultLine in DaqifiStreamingDevice
-                // had the same bug). Match `ERROR` only when followed
-                // by an SCPI delimiter (':', ' ', '!', '\t') or end of
-                // line so ordinary filenames pass through.
-                if (IsErrorResponseLine(path))
+                // Skip SCPI error responses ("**ERROR: -200, ...", "ERROR: -200, ...")
+                // and firmware status text ("Error !! ..."). Classification rule lives
+                // in ScpiResponseClassifier so it stays consistent with
+                // DaqifiStreamingDevice.IsNonResultLine (closes #190).
+                if (ScpiResponseClassifier.IsErrorResponseLine(path))
                 {
                     continue;
                 }
@@ -114,50 +109,6 @@ namespace Daqifi.Core.Device.SdCard
             }
 
             return null;
-        }
-
-        /// <summary>
-        /// Returns true if the line is a non-result error/status line that
-        /// SD listing parsers should drop. Matches both SCPI error responses
-        /// (canonical <c>**ERROR</c> marker, bare <c>ERROR</c> token followed
-        /// by a SCPI delimiter <c>:</c> / space / tab / end-of-line) and
-        /// firmware status text (<c>Error !! ...</c> with space, or the
-        /// no-space <c>Error!!</c> form). A double-<c>!</c> is required when
-        /// no other delimiter is present so legitimate filenames like
-        /// <c>error!log.bin</c> aren't dropped — single <c>!</c> alone is
-        /// ambiguous between error-status and filename. Trims both ends so
-        /// a bare <c>"ERROR\r"</c> from CRLF line endings still classifies.
-        /// Plain filenames whose basename starts with <c>error</c> /
-        /// <c>Errors</c> pass through unmatched (closes #190).
-        /// </summary>
-        /// <remarks>
-        /// Shared with <c>DaqifiStreamingDevice.IsNonResultLine</c> so any
-        /// future delimiter / trim refinement stays consistent across both
-        /// SD-response classification paths.
-        /// </remarks>
-        internal static bool IsErrorResponseLine(string line)
-        {
-            var trimmed = line.Trim();
-            return MatchesErrorPrefix(trimmed, "**ERROR")
-                   || MatchesErrorPrefix(trimmed, "ERROR");
-        }
-
-        private static bool MatchesErrorPrefix(string trimmed, string prefix)
-        {
-            if (trimmed.Length < prefix.Length
-                || !trimmed.StartsWith(prefix, StringComparison.OrdinalIgnoreCase))
-                return false;
-            if (trimmed.Length == prefix.Length)
-                return true;
-            var next = trimmed[prefix.Length];
-            if (next == ':' || next == ' ' || next == '\t')
-                return true;
-            // Single '!' is ambiguous (could be a filename like "error!log.bin").
-            // Require '!!' so we still catch firmware "Error!!" status text but
-            // let plain filenames pass through.
-            return next == '!'
-                && trimmed.Length > prefix.Length + 1
-                && trimmed[prefix.Length + 1] == '!';
         }
     }
 }
